Add tooling to run e2e tests
Link the tests dir from each sample app into test/e2e/<sample_app> so
it's easily available to edit, and a Rake task will be generated to
automatically run them when you run `rake e2e` (which is also run as
part of the default task).

This also sets up tasks to automatically run `bundle` on the gem and all
sample apps whenever we run `rake build`, to avoid pushing a version
without updating the Gemfile.lock, which results in broken tests. 😅