Run e2e tests on GitHub Actions
Because of how separate Gemfiles are required for each sample app, and
due to how ruby/setup-ruby is laid out, we require a globally-set
BUNDLE_GEMFILE to use separate gemfiles in the action.

This in turn means we need separate jobs to run each set of e2e tests,
which will slow down builds, but hopefully not make them prohibitive.

Finally, this sets up MySQL to run on the e2e container. We'll need to
break this up further as we add tests against different services, but
let's fix one thing at a time.