Add unit tests for all the sinks
This verifies that all the DB sinks we set up actually call the SQL
injection scanner when making queries with all the methods that are
exposed by the underlying library.

In some cases, like MySQL drivers, which only expose one method to
make queries, the end to end tests are enough to have coverage for
the entire sink. But SQLite and PostgreSQL provide several methods
to query the DB, and the end to end tests only cover one of those.