Poll for settings updated in a background thread
When the Agent is initialized, we start a timer that polls the runtime
API for settings updates. if the runtime API returns a fresher timestamp
than what we got on our settings, it then requests the new settings and
updates the object.

This introduces concurrent-ruby to handle thread safety, via its
TimerTask, which adds several safety guarantees, specially around
exceptions raised (useful for a background thread that will make
HTTP requests), and also allows us to use MutableStruct for the
Settings object, making it safe to update while threads are trying.
to read it.