Introduce a proper Sink abstraction
Sinks handle registering a list of scanners to run, and provide a way to
run them and track the results to report to the agent.

Their output is a "Scan" object that keeps some statistic, which can
then be aggregated at the Agent level, and possibly an "Attack" object
that can track information about a specific type of attack.