Introduce a Stats class to funnel usage stats
This is meant to be accessed by multiple threads at the same time,
mutating its state by incrementing counters and/or reshuffling the
collection of timing metrics when compressing.

Thus, this relies on LockableObject to perform all state access (reads
and writes) behind a Mutex, which is a bit of a heavy-handed approach,
but will come in handy when we implement the heartbeat event, as we need
to generate a JSON representation of the data and reset it while other
threads might be trying to access it.

Hopefully compressing timing statistics on a live production app is not
slow enough that locking around this event becomes a problem (though my
spidey sense tells me it might be for particularly big applications—we
will need to measure this).