Track the operation in which an attack happened
This keeps a Sink#operation (which defaults to the sink's name), so that
we can specify what we're patching. Also, in cases where a single sink
might patch several related methods (like with the SQL sinks), this
gives the flexibility to be specialized by scan parameters. As long as
the Attack returned from the scanner has the correct "operation" string
set, we don't need to care.