Introduce the idea of Routers
These objects take a request and map them to a Route, which contains
information about the current framework route being requested. Within
Rails, we use the framework's own Router to introspect the actual
route definition being requested.

Outside of Rails, we provide a heuristic best-guess router to map a
request URL to some form of parameterized URL that can be used in the
UI to group similar endpoints.