Simplify the Stats interface
Instead of "serialize_and_reset", just do a plain "reset" that returns a
copy of the data structures from before the reset.

With Stats tracking more structures (such as the list of requested
routes, and soon the list of outbound domains), returning the serialized
forms was becoming unwieldy.

This keeps all the concurrency operations localized in the same object,
while giving us more flexibility as to how to consume the data once the
state is reset.