Avoid a ton of errors in tests
Due to how the agent works, whenever you call #track_*, it will
implicitly start the runner (because it wants to make sure that
there is an initialized runner ready when running live).

However, starting the runner has the side effect of dispatching
the "started" event, which makes an HTTP request to our servers
which in turn gets flagged by WebMock, because this request has
not been stubbed.

Because of all of these side effects, we want the Agent methods
that perform tracking to be "disabled" during tests that result
in scans performed as a side effect, like making HTTP calls.