Add an API to track users
Developers can now call Aikido::Agent.track_user(user) from their code
whenever a User either signs in or is set as the request's current user
(i.e. by reading a cookie, or an Authorization header).

In order to keep the API simple for devs, we'll do a best guess to map
the object given to an "actor":

* If the user is an ActiveModel-like object (i.e. implements #to_model),
  we'll read the `id` and `name` attributes (configurable) from the
  model directly.

* If they pass a Hash, we'll expect either :id/:name or "id"/"name" to
  be keys, and use those values.

* If they pass an object that implements `#to_aikido_actor`, we'll call
  that and expect that the method returns an `Aikido::Agent::Actor`. The
  idea is to give users a simple protocol they can implement when they
  have complex requirements.

The last case is meant for situations where simply mapping the ID can
prove insufficient, like when multiple different types of models are
used to authenticate—for example, a `User` and `Admin`, which could have
clashing IDs (i.e. both have autoincrementing numeric IDs).

This lets users do something like

``` ruby
class Admin < ApplicationRecord
  def to_aikido_actor
    Aikido::Agent::Actor.new(id: "admin:#{id}", name: email)
  end
end
```

So they see in the UI the ID "1" and know it's the user, and the ID
"admin:1" and know it's the admin.