Detect SSRFs in the em-http-request sink
This changes the sink to use send_request which is the last possible
method we can hook to before the socket is actually opened. In
particular, this method is called _after_ all the request middleware
have been processed, and so it makes sure that any modification on
the request that happened in middleware is accounted to before scanning.

For response / redirect tracking, since all we care about is to know
that it happened, hooking to the response through a middleware is
enough, and so we use that.