Make sure we're only scanning for "private" IPs
Or more accurately, any "special use" IP.

The idea is that although we _could_ stop everything that is probably an
SSRF, we might be blocking legitimate, harmless requests by doing so.
For example, if there's a <select> on your site to let the user choose
which service to search with, including values like "www.google.com",
"bing.com", or "altavista.com", then yes, when we make a request to
"altavista.com" it _is_ very strictly speaking an SSRF.

But it's also sanitized input, and not an attack. So we can't blindly
stop every request, or we'd be getting lots of false positives and lots
of angry Zen users.

The idea is that the SSRF is dangerous if it's hitting a "private" IP,
as in, one that's likely within your internal network, and not
accessible through the internet. We consider these to be any "true"
private address from RFC1918 (10.0.0.0/8, 192.168.0.0/16, 172.16.0.0/12)
or from RFC4193 (fc00::/7), or any of the addresses defined in RFC5735
as "special use".

(Which, admittedly, you shouldn't be using in your internal network, but
we might as well be safe)