Catch SSRF attacks during DNS resolution
We can't enforce that every request to a user-supplied domain is
rejected as an SSRF, since that would cause too many false positives.

This limits the usefulness of our SSRF scanner as-is, since it will
only catch obvious attempts, like passing "localhost" or IP addresses.

Attackers are craftier than that.

This makes it so after we resolve DNS, we re-run the SSRF scanner, but
now with the knowledge of the _actual_ IP the hostname resolves to, so
we can *definitely* block requests to internal/dangerous hosts before
they happen.