Break down and refactor the Agent
The Agent was getting a little big, and it was holding a few too many
responsibilities, which made it hard to reason about, and harder to
debug problems (specially thread-related ones).

This breaks it down so:

* A "Collector" handles collecting all the statistics and runtime
  metrics that we gather.
* A "HeartbeatsManager" handles wrapping the logic for when to schedule
  the heartbeat events.
* The "Worker" deals with offloading tasks to a background thread, and
  running the delayed and recurring jobs like the heartbeat.

With all this, the Agent's sole responsibility now is coordinating these
other objects, and handling initialization and cleanup logic at shutdown.