Remove more responsibilities from the Agent
This points the top-level tracking methods directly to the Collector,
now that it is a top-level concern, removing that awkward indirection
of having to call the agent which in turn called the collector.

Simpler, and with a nicer API that is public (code wise, not API wise)
which means we don't have to fiddle with instance_variable_get-ing it
in any tests that needs to check collection state.