Do not use a middleware to rate limit in Rails
While the middleware works well for Rack apps, in Rails users are
expected to call `Aikido::Zen.track_user` in a before_action, and
so the rate limiter will never know who the current user is.

This is important when multiple users are behind the same IP (eg.
in an office), where we should use the user ID as a discriminator
between requests to determine if they should be rate limited.

This sets up a "sink" that hooks into ActionController, and adds
itself at the end of the callbacks chain (right before the action
is processed). This way we allow host apps to add the tracking call
at any point during the controller callbacks.