Add explicit protect
No sinks are loaded and the Aikido Agent does not start until the
`Aikido::Zen.protect!` method is called. This is intended to make
it safe to require `aikido-zen`. Requiring `aikido-zen` no longer
implies enabling it. Until explicitly enabled `aikido-zen` should
not change the behavior of programs that require it.