Add GET /cats/count endpoint
A new cat is added each iteration.

The GET /cats endpoint returns all cats rendered as HTML or JSON.
The size and time of the response grows with the number of cats.

The GET /cats/count endpoint returns the number of cats rendered
as HTML or JSON. The size and time of the response remains small.